requirements.txt
streamlit==1.22.0
opencv-python-headless==4.5.3.56
numpy>=1.20.0,<1.24.0
tensorflow>=2.10.0
Pillow==8.3.2
streamlit-webrtc>=0.45.0
aiortc>=1.3.0